requirements.txt
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.20
boto==2.38.0
bz2file==0.98
celery==3.1.18
chardet==2.3.0
cssselect==0.9.1
Django==1.8.2
django-celery==3.1.16
django-crispy-forms==1.4.0
django-registration-redux==1.2
feedfinder2==0.0.1
feedparser==5.1.3
jieba==0.35
jieba3k==0.35.1
kombu==3.0.26
langid==1.1.4.dev0
lxml==3.3.5
matplotlib==1.4.3
mysqlclient==1.3.6
nameparser==0.3.4
newspaper==0.0.9.8
newspaper3k==0.1.5
nltk==3.0.3
nose==1.3.7
numpy==1.9.2
oauthlib==0.7.2
pandas==0.16.2
Pillow==2.6.1
praw==3.0.0
prettyprint==0.1.5
psycopg2==2.6.1
pycparser==2.14
PyMySQL==0.6.6
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
smart-open==1.2.1
textblob==0.9.1
textblob-aptagger==0.2.0
tldextract==1.5.1
tweepy==3.3.0
update-checker==0.11
